requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
et-xmlfile==1.1.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
promise==2.3
pycodestyle==2.10.0
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
